Add loaders submodule to improve docs
